putting registration required alert before opening My Signals
