+++ app/src/main/java/org/helpapaw/helpapaw/mysignals/MySignalsFragment.java
@@ -1,7 +1,5 @@
 package org.helpapaw.helpapaw.mysignals;
 
-import android.content.DialogInterface;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -13,15 +11,12 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.databinding.DataBindingUtil;
-import androidx.fragment.app.FragmentActivity;
 
 import com.google.android.material.snackbar.Snackbar;
 
 import org.helpapaw.helpapaw.R;
-import org.helpapaw.helpapaw.authentication.AuthenticationActivity;
 import org.helpapaw.helpapaw.base.BaseFragment;
 import org.helpapaw.helpapaw.base.Presenter;
 import org.helpapaw.helpapaw.data.models.Signal;
@@ -70,7 +65,6 @@ public class MySignalsFragment extends BaseFragment implements MySignalsContract
         mySignalsPresenter.setView(this);
 
         actionsListener = mySignalsPresenter;
-        actionsListener.onOpenMySignalsScreen();
         actionsListener.onLoadMySignals();
 
         return binding.getRoot();
@@ -87,31 +81,6 @@ public class MySignalsFragment extends BaseFragment implements MySignalsContract
         super.onResume();
 
         actionsListener.onLoadMySignals();
-    }
-
-    @Override
-    public void showRegistrationRequiredAlert() {
-        final FragmentActivity activity = getActivity();
-        if (activity == null) return;
-
-        AlertDialog.Builder alertBuilder = new AlertDialog.Builder(activity)
-                .setTitle(R.string.txt_registration_required)
-                .setMessage(R.string.txt_only_registered_users_can_see_my_signals)
-                .setCancelable(false)
-                .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialogInterface, int i) {
-                        openLoginScreen();
-                    }
-                })
-                .setNegativeButton(android.R.string.cancel, null);
-        alertBuilder.create().show();
-    }
-
-    @Override
-    public void openLoginScreen() {
-        Intent intent = new Intent(getContext(), AuthenticationActivity.class);
-        startActivity(intent);
     }
 
     @Override
